refactor: refactor engine helpers into dedicated modules
