+++ src/components/ResourcePanel.tsx
@@ -1,7 +1,8 @@
 import { useGameStore } from '@/store/gameStore'
 import { GAME_TICK_INTERVAL_MS } from '@/engine/constants'
 import { calculateResourceLimits } from '@/engine/gameLoop'
-import { getAssignedCount, getPopulationCount, RESOURCES } from '@/engine/types'
+import { RESOURCES } from '@/engine/types'
+import { getAssignedCount, getPopulationCount } from '@/engine/dogs'
 import { Badge } from '@/components/ui/badge'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { ScrollArea } from '@/components/ui/scroll-area'
